Added startup class
