+++ src/API/Startup.cs
@@ -1,0 +1,40 @@
+﻿namespace API
+{
+    public static class Startup
+    {
+        public static WebApplication InitializeApp(string[] args)
+        {
+            var builder = WebApplication.CreateBuilder(args);
+            ConfigureServices(builder);
+
+            var app = builder.Build();
+            Configure(app);
+
+            return app;
+        }
+
+        private static void ConfigureServices(WebApplicationBuilder builder)
+        {
+            builder.Services.AddControllers();
+
+            builder.Services.AddEndpointsApiExplorer();
+
+            builder.Services.AddSwaggerGen();
+        }
+
+        private static void Configure(WebApplication app)
+        {
+            if (app.Environment.IsDevelopment())
+            {
+                app.UseSwagger();
+                app.UseSwaggerUI();
+            }
+
+            app.UseHttpsRedirection();
+
+            app.UseAuthorization();
+
+            app.MapControllers();
+        }
+    }
+}
